Adding support for redirect after successful authentication using ID-porten.
