+++ src/main/java/no/difi/dc2017/norgeProfil/config/SecurityConfig.java
@@ -1,0 +1,39 @@
+package no.difi.dc2017.norgeProfil.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
+import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+
+import javax.servlet.Filter;
+import javax.xml.bind.JAXBException;
+import javax.xml.transform.TransformerException;
+
+
+@SpringBootApplication
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    private Filter idportenFilter;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .antMatcher("/**").authorizeRequests()
+                .antMatchers("/", "/login**", "/webjars/**", "/built/**", "/img/**", "/styles/**","/fonts/**", "/user", "/data/**", "/norge", "/sertifikat", "/reservasjon").permitAll()
+                .anyRequest().authenticated()
+                .and()
+                .exceptionHandling().authenticationEntryPoint(new LoginUrlAuthenticationEntryPoint("/"))
+                .and()
+                .logout().logoutRequestMatcher(new AntPathRequestMatcher("/logout")).logoutSuccessUrl("/").permitAll()
+                .and()
+                .csrf().csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
+                .and()
+                .addFilterBefore(idportenFilter, BasicAuthenticationFilter.class);
+    }
+}
