+++ src/main/java/no/difi/dc2017/norgeProfil/web/UserController.java
@@ -1,0 +1,20 @@
+package no.difi.dc2017.norgeProfil.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import java.security.Principal;
+
+/**
+ * @author erlend
+ */
+@Controller
+@RequestMapping("/user")
+public class UserController {
+
+    @RequestMapping(method = RequestMethod.GET)
+    public Principal user(Principal principal) {
+        return principal;
+    }
+}
