+++ src/main/resources/application.properties
@@ -1,4 +1,3 @@
-
 server.port=8080
 spring.thymeleaf.mode: LEGACYHTML5
 #spring.mvc.view.prefix=/WEB-INF/views/
